search within channel
